NFC: Consolidate all render configuration in default_configs.rs
Dump render config as well as normal render config was spread out across
the codebase. This commit moves code around to consolidate it all in the
same place. This not only makes code easier to discover but solves
dependency cycle issues for the next commit.
