+++ below/dump/src/cgroup.rs
@@ -13,108 +13,8 @@
 // limitations under the License.
 
 use model::SingleCgroupModelFieldId;
-use render::RenderConfig;
 
 use super::*;
-
-impl HasRenderConfigForDump for model::SingleCgroupModel {
-    fn get_render_config_for_dump(field_id: &SingleCgroupModelFieldId) -> RenderConfig {
-        use common::util::get_prefix;
-        use model::CgroupCpuModelFieldId::ThrottledPct;
-        use model::CgroupIoModelFieldId::CostIndebtPct;
-        use model::CgroupIoModelFieldId::CostIndelayPct;
-        use model::CgroupIoModelFieldId::CostUsagePct;
-        use model::CgroupIoModelFieldId::CostWaitPct;
-        use model::CgroupIoModelFieldId::DbytesPerSec;
-        use model::CgroupIoModelFieldId::DiosPerSec;
-        use model::CgroupIoModelFieldId::RbytesPerSec;
-        use model::CgroupIoModelFieldId::RiosPerSec;
-        use model::CgroupIoModelFieldId::RwbytesPerSec;
-        use model::CgroupIoModelFieldId::WbytesPerSec;
-        use model::CgroupIoModelFieldId::WiosPerSec;
-        use model::CgroupMemoryModelFieldId::Anon;
-        use model::CgroupMemoryModelFieldId::File;
-        use model::CgroupMemoryModelFieldId::Pgactivate;
-        use model::CgroupMemoryModelFieldId::Pgdeactivate;
-        use model::CgroupMemoryModelFieldId::Pgfault;
-        use model::CgroupMemoryModelFieldId::Pglazyfree;
-        use model::CgroupMemoryModelFieldId::Pglazyfreed;
-        use model::CgroupMemoryModelFieldId::Pgmajfault;
-        use model::CgroupMemoryModelFieldId::Pgrefill;
-        use model::CgroupMemoryModelFieldId::Pgscan;
-        use model::CgroupMemoryModelFieldId::Pgsteal;
-        use model::CgroupMemoryModelFieldId::Shmem;
-        use model::CgroupMemoryModelFieldId::Slab;
-        use model::CgroupMemoryModelFieldId::Sock;
-        use model::CgroupMemoryModelFieldId::Swap;
-        use model::CgroupMemoryModelFieldId::ThpCollapseAlloc;
-        use model::CgroupMemoryModelFieldId::ThpFaultAlloc;
-        use model::CgroupMemoryModelFieldId::Total;
-        use model::CgroupMemoryModelFieldId::WorkingsetActivateAnon;
-        use model::CgroupMemoryModelFieldId::WorkingsetActivateFile;
-        use model::CgroupMemoryModelFieldId::WorkingsetNodereclaim;
-        use model::CgroupMemoryModelFieldId::WorkingsetRefaultAnon;
-        use model::CgroupMemoryModelFieldId::WorkingsetRefaultFile;
-        use model::CgroupMemoryModelFieldId::WorkingsetRestoreAnon;
-        use model::CgroupMemoryModelFieldId::WorkingsetRestoreFile;
-        use model::CgroupMemoryModelFieldId::Zswap;
-        use model::CgroupPressureModelFieldId::MemoryFullPct;
-        use model::CgroupPressureModelFieldId::MemorySomePct;
-        use model::SingleCgroupModelFieldId::Cpu;
-        use model::SingleCgroupModelFieldId::Io;
-        use model::SingleCgroupModelFieldId::Mem;
-        use model::SingleCgroupModelFieldId::Name;
-        use model::SingleCgroupModelFieldId::Pressure;
-        use render::HasRenderConfig;
-
-        let rc = model::SingleCgroupModel::get_render_config_builder(field_id);
-        match field_id {
-            Name => rc.indented_prefix(get_prefix(false)),
-            Cpu(ThrottledPct) => rc.title("Throttled Pct"),
-            Io(RbytesPerSec) => rc.title("RBytes"),
-            Io(WbytesPerSec) => rc.title("WBytes"),
-            Io(DbytesPerSec) => rc.title("DBytes"),
-            Io(RiosPerSec) => rc.title("R I/O"),
-            Io(WiosPerSec) => rc.title("W I/O"),
-            Io(DiosPerSec) => rc.title("D I/O"),
-            Io(RwbytesPerSec) => rc.title("RW Total"),
-            Io(CostUsagePct) => rc.title("Cost Usage"),
-            Io(CostWaitPct) => rc.title("Cost Wait"),
-            Io(CostIndebtPct) => rc.title("Cost Indebt"),
-            Io(CostIndelayPct) => rc.title("Cost Indelay"),
-            Mem(Total) => rc.title("Mem Total"),
-            Mem(Swap) => rc.title("Mem Swap"),
-            Mem(Zswap) => rc.title("Mem Zswap"),
-            Mem(Anon) => rc.title("Mem Anon"),
-            Mem(File) => rc.title("Mem File"),
-            Mem(Slab) => rc.title("Mem Slab"),
-            Mem(Sock) => rc.title("Mem Sock"),
-            Mem(Shmem) => rc.title("Mem Shmem"),
-            Mem(Pgfault) => rc.title("Pgfault"),
-            Mem(Pgmajfault) => rc.title("Pgmajfault"),
-            Mem(WorkingsetRefaultAnon) => rc.title("Workingset Refault Anon"),
-            Mem(WorkingsetRefaultFile) => rc.title("Workingset Refault File"),
-            Mem(WorkingsetActivateAnon) => rc.title("Workingset Activate Anon"),
-            Mem(WorkingsetActivateFile) => rc.title("Workingset Activate File"),
-            Mem(WorkingsetRestoreAnon) => rc.title("Workingset Restore Anon"),
-            Mem(WorkingsetRestoreFile) => rc.title("Workingset Restore File"),
-            Mem(WorkingsetNodereclaim) => rc.title("Workingset Nodereclaim"),
-            Mem(Pgrefill) => rc.title("Pgrefill"),
-            Mem(Pgscan) => rc.title("Pgscan"),
-            Mem(Pgsteal) => rc.title("Pgsteal"),
-            Mem(Pgactivate) => rc.title("Pgactivate"),
-            Mem(Pgdeactivate) => rc.title("Pgdeactivate"),
-            Mem(Pglazyfree) => rc.title("Pglazyfree"),
-            Mem(Pglazyfreed) => rc.title("Pglazyfreed"),
-            Mem(ThpFaultAlloc) => rc.title("THP Fault Alloc"),
-            Mem(ThpCollapseAlloc) => rc.title("THP Collapse Alloc"),
-            Pressure(MemorySomePct) => rc.title("Mem Some Pressure"),
-            Pressure(MemoryFullPct) => rc.title("Mem Pressure"),
-            _ => rc,
-        }
-        .get()
-    }
-}
 
 pub struct Cgroup {
     opts: GeneralOpt,
