+++ below/render/src/lib.rs
@@ -397,6 +397,25 @@ pub trait HasRenderConfig: Queriable {
     }
 }
 
+pub trait HasRenderConfigForDump: HasRenderConfig {
+    fn get_render_config_for_dump(field_id: &Self::FieldId) -> RenderConfig {
+        Self::get_render_config(field_id)
+    }
+
+    /// Configures how to dump model fields in OpenMetrics format
+    ///
+    /// Some fields cannot be dumped in openmetrics format, for example strings. For those,
+    /// return None.
+    fn get_openmetrics_config_for_dump(
+        &self,
+        _field_id: &Self::FieldId,
+    ) -> Option<RenderOpenMetricsConfigBuilder> {
+        // XXX: the default implementation will be deleted in the final commit. We add one
+        // here to make each commit compilable.
+        None
+    }
+}
+
 #[test]
 fn test_openmetrics_gauge() {
     let config = RenderOpenMetricsConfigBuilder::new(OpenMetricsType::Gauge)
